+++ D&Cv2/tempCodeRunnerFile.cpp
@@ -1,0 +1,2 @@
+
+  SisInventario inv1;
