--- DInosYCalabozosV2/prueba.bin
@@ -1,1 +1,0 @@
-0x480x650x6C0x6C0x6F0x2C0x200x570x6F0x720x6C0x640x2
