--- TestSite.cpp
@@ -1,23 +1,0 @@
-#include "PrototipoPersonaje.h"
-#include "ExpJugador.h"
-#include "Personaje.h"
-#include <iostream>
-#include <vector>
-#include <string> 
-#include <locale>
-// archivo: ExecuteTest.cpp a 5/30/2024: TestSite.cpp
-// autor: Daniel Emiliano Lopez Aguilar
-//  fecha: 5/27/2024
-/* descripción:
-Este archivo funciona como zona de pruebas para snippets de codigo
- y dudas sobre la POB en C++ con la intencion de no alterar el
- funcionanmiento de los archivos.
-*/
-
-
-int main(int argc, char const *argv[])
-{
-  Espadachin tomCruise;
-  tomCruise.ataqueMordidaFinal();
-  return 0;
-}
