--- TestSite2.cpp
@@ -1,9 +1,0 @@
-  double ataque_flechaRapida(){
-    const string text_array[] = {
-  "\\\\\\\\\\\\_____________________\"-._",
-  "        /////~~~~~~~~~~~~~~~~~~~~~/.-'"
-};
-  
-
-   Personaje(string nombre,long int vida, long int stamina, long int ataque, long int magia, long int defensa, long int agil)
-        :m_NOMBRE(nombre), m_vida(vida), m_stamina(stamina), m_ataque(ataque), m_magia(magia), m_defenza(defensa), m_agilidad(agil){};
