+++ todos.txt
@@ -1,66 +1,2 @@
-Beto arbol de dialogos 4 finales. 
-sis de dialogos
-
-1:40 pm viernes 5/31/2024
-2 horas de trabajo activo mas o menos.
-  metodo main 
-
-se puede ejecutar un cpp por medio de otro cpp? si, pero require metodos especiales, no es recomendable.
-para que sirven los archivos .h? para ser usados como "interfaz" se pueden ejecutar en otros cpps y h 
-
-
-yo sistema de combate (return expGanada, karma);
-  protopersonaje a personaje y enemigo.
-  Agregar metodo de quitar inventario a protopersonaje.
-  instancia de enemigos.
-  textos en sistema
-  implemenar los dibujos ascii.
-  DONE dibujos asccii incluir numeros del 0 al 9
-  dibujos ascii incluir mensajes extremos
-  pasar los que necesiten sizeof():
-            #include <array>
-
-          std::array<int, N> arr; // N es el tamaño del arreglo
-
-          // Obtener el tamaño del arreglo
-          size_t size = arr.size();  
-
-
-
-          
-  clase interfazDeUsuario()
-
-  crear listaDeObjetosParaInventario{
-
-        double m_
-        double m_vida;
-        double m_stamina;
-        double m_ataque;
-        double m_magia;
-        double m_defenza;
-        double m_agilidad;
-  }
-
-  movimiento lineal en base de encuentros 
-    movimiento lineael continuo en avance a eventos aleatorios de dos tipos: comunes, unicos y canonicos.
-
-    que es un evento canonico:
-      estos cambian las ramas de historias.
-
-  
-
-
-
-
-imprimir digitos
-
-
-6/8/2024
-puedo notar que la clase Clase y Personaje son iguales, voy a mover 
-las funciones de Clase a Personaje antes conocido como protoPersonaje
-
-asi las clases de enemigo y de jugador, se derivan de las clases Enemigo y Jugador en vez de
-clase. 
-
-Creo que es irrelevante las clases guerrero, magico sigilo, esto ya no es un piedra, papel y tijeras.
+8/18/2024: Todos los metodos de inventario sirven :D
